+++ src/rc/render.js
@@ -1,16 +1,8 @@
-import uniqid from 'uniqid';
-
-import { Store, module_identity, propagate_transducer, attach_transducer } from './state';
-
-const compose = (...func) => (...args) => func.reduce((acc, cur) => cur(acc), ...args);
-const pipe = (...func) => compose(...func.reverse());
-
+import { Store } from './state';
 
 export const anchor = (mithril) => ({
     'oninit': function({ 'attrs': { store, provider }}){
-        console.log('anchor for ', store.getState(), store.getChain());
         this.component = provider.component(store);
-
     },
     'view': (vnode) => {
         return mithril(
@@ -22,7 +14,6 @@ export const anchor = (mithril) => ({
         );
     }
 });
-
 
 export const anchor_group = (mithril) => ({
     'oninit': function({ 'attrs': { provider }}){
@@ -47,7 +38,7 @@ export const anchor_group = (mithril) => ({
 });
 
 
-const component_transducer = (filter_resource) => (item) => (next) => (store = null) => {
+export const component_transducer = (filter_resource) => (item) => (next) => (store = null) => {
         const next_component = next(store);
         if(store === null || filter_resource !== store.getState().resource) return next_component;
 
@@ -58,7 +49,7 @@ const component_transducer = (filter_resource) => (item) => (next) => (store = n
     }
 ;
 
-const controller_transducer = (filter_resource) => (provider, action_creators = null) => (next) => (store) => {
+export const controller_transducer = (filter_resource) => (provider, action_creators = null) => (next) => (store) => {
     const next_component = next(store);
     if(store === null || filter_resource !== store.getState().resource) return next_component;
 
@@ -76,51 +67,3 @@ const controller_transducer = (filter_resource) => (provider, action_creators = 
         }
     };
 };
-
-export class Provider {
-    constructor(mithril){
-        this.component = (/*query_store*/) => ({ 'view': () => mithril('#') });
-        this.reducer = (/*query_resource*/) => (state) => state;
-        this.Anchor = anchor(mithril);
-        this.AnchorGroup = anchor_group(mithril);
-    }
-
-    connect_component(resource, component, action_creators){
-        this.component = compose(
-            component_transducer(resource)(component),
-            controller_transducer(resource)(this, action_creators)
-        )(this.component);
-    }
-
-    connect_transducers(transducers){
-        this.reducer = pipe(
-            ...transducers
-        )(this.reducer);
-    }
-}
-
-export const draft = {
-    Provider,
-    module_identity,
-    propagate_transducer,
-    attach_transducer
-};
-
-
-const renderer_action_creators = (id, chain) => ({
-    'attach': ({ resource, consumer_state }) => ({
-        'type': 'ATTACH_MODULE',
-        resource,
-        consumer_state,
-        'parent': id,
-        'chain': [...chain, this.id]
-    })
-});
-
-/**
-new provider
- + initial_state => new store
-    + resource, component, action_creators, transducers => new resource
-        render
-
-**/
